codegen missing constants used in annotations
Summary:
Handled not stubbed constants in annotations parameters.
Example:

import com.instagram.direct.realtime.constants.DirectRealtimeConstants.THREAD_TAG

ThreadConfined(THREAD_TAG)

 private fun transitionTo() {

    …

 }

In this example we do not stub THREAD_TAG, because it does not contribute to ABI. But with the new Kosabi design, the backend tries to inline the annotation’s value and fails with an exception.

K1 jvm-abi-gen was just ignoring unresolved constant values and was generating ABI with no-parameters in the annotations.

Differential Revision: D86302853

fbshipit-source-id: ca8f984699e5655dbd002f1450b093f7e5ab00e5