Create classes subtarget
Summary:
### Context
I am trying to minimize the output size generated by KSIC to decrease Buck memory usage associated with directory structure processing.

### This stack
The goal of this stack is to eliminate the duplicated `__classes__` directory in `incremental_state` and instead utilize the `__classes__` output. To achieve this, the Kotlinc output must be isolated by creating a separate folder: `__kotlin_classes__`.

### This diff
This diff creates an `kotlin_classes` subtarget to be able to expose the dir in e2e tests.

Reviewed By: NavidQar

Differential Revision: D86412881

fbshipit-source-id: c42711d726ca5d48a9338b100a9c8ac155a21113