Add hardcode permissions for deterministic output flag to android_aar rule
Summary: D67449762 removed the option to hardcode AAR file permissions from the AAR builder binary, because it was unused. However, I now have a use case for a skycastle workflow to check if a AAR target needs rebuilding. Add a flag `hardcode_permissions_for_deterministic_output` to the `android_aar` prelude rule to remain consistent with the other sites where a similar flag is used.

Reviewed By: christolliday

Differential Revision: D86701335

fbshipit-source-id: 259890f872dfb2f15c0b0630039ba4d9482caeb5