buck2: Add modifier support to check_dependencies_test and make target_deps() filter optional
Summary:
I'd like to assert that when a target is built with specific modifiers, it doesn't bring in other (incorrect) versions of third-party libraries. This adds modifiers support to this test by passing along the appropriate modifier args to the bxl command.

This also makes the `target_deps()` filter optional, something that the older `audit_dependencies_test` supported.

Reviewed By: scottcao

Differential Revision: D86464757

fbshipit-source-id: c66a5f94bbd8082569210f0d0d105b0c69ef47f8