util: Add `async_move_clone`
Summary:
See some of the discussion in D86390097, makes this pattern less confusing

I'm happy to take feedback on whether this is actually an improvement. It's more explicit, but also macros are scary and complicated, idrk.

Used this as an opportunity to learn a little bit of ast-grep to find places to apply this:

```
../../tools/third-party/ast-grep/ast-grep run --globs 'app/*/src/**/*.rs' --pattern '{ let $V = $V.clone(); async move { $$$X }}' --rewrite 'buck2_util::async_move_clone!($V, { $$$X })' -U
```

Looks like a neat tool.

There are many more uses of this pattern that are just less explicit, would need to learn a bit more about ast-grep to find those ig

Reviewed By: Will-MingLun-Li

Differential Revision: D86681945

fbshipit-source-id: 43a7d390871600e24180599de3d8800cdb081ba9