cgroups: Ensure forward progress in kill & retry case
Summary:
See the added comment - previously it was possible that we might kill and retry all actions and never let any of them finish.

The previous choice of suspending the largest action is also specifically pretty wrong in the kill & retry case; the largest action is probably the one that represents the most completed useful work, so by killing that one we throw away lots of work. This implementation instead prefers to kill recently spawned actions, which have probably done much less work

In the freeze case it made more sense, but wasn't load bearing; after all, all actions do have to eventually finish, and maybe running the largest one first is actually better as that way if an OOM was inevitable, it can happen sooner

A further reason to do this is that we may one day actually start being non-trivially intelligent about the order in which we schedule local actions; in that case, preserving the order seems good as a default

Reviewed By: Will-MingLun-Li

Differential Revision: D86713953

fbshipit-source-id: 5c37ef2f86971fa4c5a64d0a0e8883bcb07d082e