cgroups: Perform action freezing in the executor
Summary:
Right now, the actual work of freezing the cgroup happens in the resource manager; this has a bunch of problems, primarily the ones documented in the new comment - we might freeze during a kill or during a spawn, and we really don't want that.

Doing the actual freezing in the same code that manages the rest of the execution fixes those bugs.

This diff replaces the "actually write to `cgroup.freeze`" logic in the resource manager with instead producing a stream of freeze events; that stream is then passed down the pipeline until it eventually ends up in the forkserver where the freezing will happen

Reviewed By: Will-MingLun-Li

Differential Revision: D86681955

fbshipit-source-id: e1daa595bfd7b49d68cffa46a5e25799b78d4bf3