Add Content Based Paths Flag to Python Toolchain Info
Summary:
## Context:

See T239185724. We want to enable content-based output path addressing for Python bytecode.

 ---

## This Diff:

Just add a boolean field to the Python toolchain to indicate whether content-based path hashing should be enabled or not.

We will pipe this through to the compile action later.

For now, it's hardcoded to False (and also is not referenced anywhere) so this is a no-op change.

Reviewed By: christycylee, IanChilds

Differential Revision: D83368271

fbshipit-source-id: 86f471cca7430b586ee801038931cfbbdab65a86