Create subtarget for recursive clang traces
Summary:
Previously there was only a target to get a clang trace
for an individual library. This creates a recursive target using tsets
to allow getting clang traces or a library and its deps. Also, expose
the same targets on cxx_binary so we can get it with top level builds.

Reviewed By: milend

Differential Revision: D86563117

fbshipit-source-id: 7f805850329dc53f22150af9044d028c726374c0