Simplify logic for handling testcases that didn't run
Summary:
# Context
When running a suite, there are various reason why a testcase may end up not running. Most likely because some previous init_per function failed, but it is also possible that the user explicitly skipped it. The test runner needs to figure out what case we are in, and mark it as either a failure or an omission.

# Problem
The logic to do this is unnecessarily convoluted. We have two functions that do the exact same thing, one calling the other for some cases. The only actual differences are:

* One of the two uses `timeout` for the testcase if the init failure was a `timeout`, but we don't emit `timeout` internally, so this is mostly theoretical
* The error messages of one of the functions are more clear than the other, but in practice, we use the less clear ones that talk about "results not being recorded" which for the user is going to be rather cryptic

# This diff
We simplify by keeping only one of the two functions, the one with the more clear error messages.

Reviewed By: TD5

Differential Revision: D87076370

fbshipit-source-id: c7e68352a2b2722b5c3f9c62ab5fc3244abfbc81