Disallow type annotations anywhere other than first declaration
Do not allow `x=5; x: str = ""`, or `x: int = ...; x: str = ...`,
or `def func(): pass; def func(): pass`. This is invalid in python,
so should not be valid starlark.

Updates a golden test, which made this error. Also removes a `def assert_eq`
in another test as the function already exists.