Type check every function and top-level scope separately
Previously, we gathered and solved bindings for every top-level def,
and not for a module's root bindings.

This was a deliberate limitation for typechecker performance reasons,
introduced in D48752028/D48752027 back in 2023.

This completes the work described in those diffs, which said ideally:

1. typecheck top-level assignments
2. then evaluate defs

This does one better and typechecks even nested defs individually. This achieves the
goal of reducing the iterations needed in solve_bindings by passing the
smallest possible quantity of bindings to solve_bindings at any one time.