Apple: enforce that we always return `target_sdk_version`
Summary:
We want to get rid of setting `min_sdk_version` on Apple `cxx_toolchain()` because it only sets the minimum _allowable_ OS version. This is not necessary as we should be setting the `target_sdk_version` for all of them anyway (i.e., minimum OS version).

We will use `fail()` to determine if this field is actually ever returned. If not, we can safely remove it.

#buildall-fbobjc

Reviewed By: rmaz

Differential Revision: D87079474

fbshipit-source-id: 1f6863114da14e63031d70d6ed5a40988ceb3bf2