add prefix map for TMPDIR
Summary:
Swift compilation with macros will embed the path to generated source code in the debug info. These paths are in the temp dir:
https://github.com/swiftlang/swift/blob/cf36d0053c5bb03d0b5732b06acca0c09046d555/lib/Basic/SourceLoc.cpp#L247-L250

Local and remote actions will have different values set for `TMPDIR`, so we need to include this in the debug path map to ensure deterministic output.

Reviewed By: benb, chatura-atapattu

Differential Revision: D87085633

fbshipit-source-id: effe114e32410b509b4835c85ad65c317a0c9637