Improve error wording
Summary:
rust-project error messages are visible to users in VS Code:

 {F1983566538}

This is most common when switching commits and the BUCK file no longer contains the Rust targets you've newly defined in your branch.

The error message is also confusing: "owning targets" makes sense when the input was a path to a source file, but not when the input was a path to a BUCK file.

Instead, show a specific error message depending on the argument that was passed, and ensure the progress message makes sense for all arguments.

Differential Revision: D87224749

fbshipit-source-id: 7513e55c02bc979a2d108212b74f9d7d0ca3821f