Use content based output paths for dthin-lto actions where possible
Summary:
Enable content based output paths for all dthin-lto actions that support it. The only case that does not support is this pre-merged bitcode files produced during thin-link. As the comment outlines, those files are referenced by file path in index summary shard files that are published by the same action, so we cannot use content based paths, and the paths would be incorrect when written into the summary shard files.

Include a new target platform with which to run dthin-lto tests.

Reviewed By: rmaz

Differential Revision: D85798762

fbshipit-source-id: 9cd297826cca8d8d51a003a9d7494b31cb100899