Drop all DICE nodes when FS events missed
Handle the case where the notify crate tells us that events were missed.  In
this case, when sync is called at the beginning of a new build, the entire DICE
graph gets dropped.  This should fix the bug where caches can become out of
sync with the file system due to missed events.