Replace SERVER_STDERR_UNKNOWN/EMPTY tags
Summary:
A big chunk of our error rate is from SERVER_STDERR_UNKNOWN errors, which should only happen if no signal failure was logged to stderr and no structured error was logged to a buckd.error file which typically means the daemon was killed externally. The only difference between SERVER_STDERR_UNKNOWN and SERVER_STDERR_EMPTY is that the latter happens before any (typically unrelated) log lines write to stderr.

They are both logged when the connection initializes (DAEMON_CONNECT) or if it disconnects later (DAEMON_DISCONNECT).
If this happens during connection it should usually be a tier0 error, if it happens later it should be an environment error.

Reviewed By: Will-MingLun-Li

Differential Revision: D87516028

fbshipit-source-id: fcc26d7d1c9b7fc09ebe6d787bcde0d86cdc026f