do not consider all classpath files as autoimported
Summary:
The current problem with SamePackageClassStubsGenerator it considers types on classpath as not requiring import, so if both another target with the same package name and a target on classpath will have the same type name, stubsgen won't stub the type, considering it to be used from the classpath (even though it's not imported).

Why it's causing problems?

First of all, it's fragile, cause the truth is that only SDK types do not require imports. It means even if we have a non-sdk type on classpath and it's not imported - we should stub it, cause it comes from the same package, if it came from the classpath - it would be imported, otherwise the main library build would fail.

The concrete example:
When we pass android.jar on classpath for so-abi builds D84510978.
The build  buck2 build "fbsource//fbandroid/java/com/facebook/qe/schema:factory[source-only-abi]" fails with an error:

unresolved reference: Schema.

The type Schema is supposed to come from a target https://www.internalfb.com/code/fbsource/[c5b1b8c09d5be8fb5a6f902406f43ff0a0d7946a]/fbandroid/java/com/facebook/qe/schema/BUCK?lines=21 with the same package name, so we should stub it. But after we start passing android.jar, stubsgen thinks that Schema refers to javax.xml.validation.Schema which is part of android.jar, even though it's not imported and has to imported explicitly, so Schema isn't stubbed anymore and abi-gen fails.

Reviewed By: siaojiecai

Differential Revision: D86866948

fbshipit-source-id: 8f8b418f51c856c0463977d85ace21181ec20976