Split 'synchronizing-and-waiting' into three categories
Summary:
When we’re looking at critical path events we lump several categories of work into the same bucket which results in us not being able to determine if there was a bad confounding factor when looking at the resulting metrics (in particular if there was a concurrent command for which we were waiting). This splits this into three subsections - one for waiting for concurrent commands, one for the FS watcher syncing, and then a catch-all. This is still a little odd because we get the catch-all by subtracting off individual times from the older E2E time but that's probably fine.

We could wrap both of these in some sort of state object to prevent there being this many arguments to exec()... but I don't know how strongly I feel about it.

Reviewed By: cjhopman

Differential Revision: D87566371

fbshipit-source-id: 56b4634f7db324236d94f4e65c56025549b86c66