Collapse Pytorch generated stubs into owning targets
Summary:
The way Pytorch generated stubs work is that they're generated with genrules, but both are depended on by and depend on dependent targets. Because of this, we should pull these generated files into the target that defines them as type stubs.

We can do this by looking at the special generated `=` target (which is disallowed almost everywhere else in buck), and using that to figure out a couple of things about the target, such as:
- what to generate,
- where the output file is
- how we should construct its module name

With this diff, we should recognize Pytorch generated files and go-to-def within them (back and forth from buck-out), except for the fact that Pytorch has its own Pyrefly config.

Reviewed By: grievejia

Differential Revision: D87904032

fbshipit-source-id: 47793392e0d91406802688b8bfc9504d49e98237