Insert progress markers in stdout
Summary:
# Context
At the moment, the logic we use to "capture" stdout in tests is extremely convoluted. The test-runner node hacks up the `user` process to intercept (some, but not all) the produced stdout, and runs some sort of running window of lines.

This has several drawbacks:
  * It is memory heavy (specially since we use strings)
  * The capture logic is not perfect and we tend to omit duplicated lines or miss lines
  * It is not currently possible to "stream" stdout as the test runs

# General solution
We want to move to a state where we instead log progress markers to stdout, and later postprocess the stdout file to find what portion corresponds to each testcase, etc.

# This diff
As a first step, we introduce the logic to log these markers, but don't process them at the moment..

Reviewed By: TheGeorge

Differential Revision: D86784737

fbshipit-source-id: d48344a0063b8f4e044a8d3feb5d40b4f6ef5d30