Let the user now when stdout was truncated
Summary:
# Context
When a testcase fails, we show the stdout limited to that testcase. However, we truncate stdout in case it is deemed too large.

# Problem
While we used to show a message to the user when this was happening, since we moved the implementation so that the stdout log is parsed, this no longer happens.

# This diff
We add a notice when stdout is truncated.

Reviewed By: TheGeorge

Differential Revision: D87359702

fbshipit-source-id: 3a2d97df1e550979e7ae5ff039638e862bfaa1f7