Remove stdout collection from test-binary
Summary:
# Context
In the previous diffs we have move the way we collect the "stdout" for a single testcae. Before, it would happen in the test-executor node, while now it is performed in the test-binary, by processing the log file of stdout.

# Problem
The code for collecting std_out on the test-executor is still there, but nothing relies on it

# This diff
Remove this logic

Reviewed By: TheGeorge

Differential Revision: D87359701

fbshipit-source-id: f981e8480bf5f482f455e360ce3b27a589922ad3