Remove bogus stdout accumulation in test-binary
Summary:
# Context
To run an Erlang suite, the test-binary opens a port to create the test-executor node that will actually run the suite and report back the results.

# Problem
At the moment, the test-executor node (tries to) capture all of its stdout and writes it to a file. However, the port loop in the test-binary assumes that it can receive some and "accumulates" it in its local state, but does nothing with it

# This diff
Remove all that accummulation logic

Differential Revision: D87633130

fbshipit-source-id: d7dc24f57f81e113a1f30533123a5441ca80a220