Process progress markers from stdout log file
Summary:
# Context
At the moment, the logic we use to "capture" stdout in tests is extremely convoluted. The test-runner node hacks up the `user` process to intercept (some, but not all) the produced stdout, and runs some sort of running window of lines.

This has several drawbacks:
  * It is memory heavy (specially since we use strings)
  * The capture logic is not perfect and we tend to omit duplicated lines or miss lines
  * It is not currently possible to "stream" stdout as the test runs

# General solution
We want to move to a state where we instead log progress markers to stdout, and later postprocess the stdout file to find what portion corresponds to each testcase, etc.

# Problem
In D86784737 we made it so that for every testcase, we emit progress markers to stdout, but were not doing anything with them

# This diff
Here we make it so that after the suite finishes running, the "test-binary" processes the stdout log file in the following way:
  * All (*) markers are removed, producing a cleaned up log file without markers
  * For each start-marker, the collected stdout is returned, capped to a certain value, in case a test-case just outputs "too much".

In the process, we now ensure that every marker starts at the beginning of the line, by inserting an extra leading newline. While processing, this extra newline will be removed (except if it was completing a line that lacked a newline)

(*) Because collection of stdout still happens in the "test-executor" node, which is currently returning duplicated lines, for each duplicate progress marker, the second one remains in the cleaned-up logs. This will be fixed later in the stack when we remove the collection of stdout from the "test-executor" node.

Reviewed By: TheGeorge

Differential Revision: D86984397

fbshipit-source-id: d431d79d51e216cfd3ebea3a7f580c316d244764