Use distinct types for test-tree results and collected results
Summary:
# Context
When running Erlang tests, the CT suite runs in a seperate "test-executor" node, and the results passed back to the "test-binary" node that reports them back to TPX.

In the "test-executor" the execution results are stored in a "test-tree", where the root corresponds to the the suite and its init_per/end_per callbacks; inner nodes to groups and leaves to testcases. The "test-binary" receives this and "collects" the results, which essentially means deciding for each "missing" testcase, whether it failed or was skipped, and changing the representation to a list of testcase results plus the list of results of all their related init_per/end_per.

# Problem
Both representations discussed above currently use the same `method_result()` type. But as we are moving to a scenario where on the test-executor we keep only "stdout progress markers" and the actual stdout is computed later in the "test-binary", we would like this type to be different: on the test-executor it will contain the markers, on the binary it will contain only the stdout blob.

# This diff
This is mostly plumbing. We introduce a new `collected_method_result()` type that doesn't contain the progress markers, and modify all the types accordingly. For now we introduce a trivial function to convert from one type to the other, that simply ignores the progress markers, but later in the stack, we'll have to enrich the functoins with the blobs of stdout for each method, and use that instead.

Reviewed By: TheGeorge

Differential Revision: D87146697

fbshipit-source-id: bb807998c5bff5b5701d0c8c4f9618a63f92cb02