Improve types of some specs
Summary:
# Context
The test-binary writes a CT test-spec to a file, that is then used to run the desired tests. The test-spec contains the `ct_hooks` configuration, and this is where the `cth_tpx` hook is injected. This hook is the one that does the test-result colllection.

# Problem
Some of the types involved in building this part are a bit weak. In particular `string()` is being used for what is supposed to be the name of the file that is to store the test results

# This diff
Use ``file:filename_all()` instead.

Reviewed By: TheGeorge

Differential Revision: D87977510

fbshipit-source-id: ca0db52a08a0a513e6fe3e14faf5170121e1768c