Improve building of CT hooks in spec
Summary:
# Context
The test-binary writes a CT test-spec to a file, that is then used to run the desired tests. The test-spec contains the `ct_hooks` configuration, and this is where the `cth_tpx` hook is injected. This hook is the one that does the test-result colllection.

# Problem
The logic to collect and merge all user-provided `{ct_hooks, ...}` entries is a bit convoluted and accidentally quadratic (as the list of user-privded options gets re-scanned from the beginning once for each `ct_hooks`

# This diff
Use `lists:splitwith()` to collect all the `ct_hooks` instead

Reviewed By: TheGeorge

Differential Revision: D87978009

fbshipit-source-id: 931b8c09e0eca35d88888f3b1290dbbf167e760c