remote_execution: remove check for exec_enabled on ExecutionCapabilities
I have no idea why this exists. It's completely fine if a server has
`exec_enabled = false`, as long as the selected platform used by Buck
does not try to perform remote execution; this is the case when you're
only using caching and not RE -- the server will correctly specify
its capabilities, but this check will cause Buck to not use it at all.
Remove it.

Signed-off-by: Austin Seipp <aseipp@pobox.com>