Populate oncall to tpx from buck.oncall attribute directly
Summary:
Right now, how oncalls are populated is strictly from the starlark codebase.  That is to say, that every test-defining macro must correctly take the package oncall and pass it through in the contacts or oncall (or both) fields.

It turns out that many macros do not do this correctly.  We have a fallback, where CI attempts to provide an oncall to tpx.  This works, unless we have targets with multiple oncalls in the same job, which we frequently do.

This code attempts to guarantee that if the oncall is provided on the BUCK file and its not provided in a closer layer (e.g. on the target itself), we provide the BUCK oncall by default.

For internal projects, we expect users to define oncalls on BUCK files and not individual targets, so this should help cover a known-gap.

That all being said, my expertise in the buck codebase is low and would appreciate expert review on the implications for buck.

Related thread that talks about the issue: https://fb.workplace.com/groups/sandcastleqna/posts/29999785469643349/?comment_id=30003849522570277

My end-goal is to reduce the number of unowned builds and tests - some of the most frequently run, but unowned tests are due to using macros that don't correctly populate the oncall.  Instead of playing whack-a-bug, I thought I could solve the issue here for all.

RFC post: https://fb.workplace.com/groups/buck2dev/posts/4189436801344360

Reviewed By: gpires-meta

Differential Revision: D87573868

fbshipit-source-id: 367924471a1bdcc6a5345cf252996b6862d8351d