Optionally make command_alias respect already set env vars
Summary: I've always found it annoying that `command_alias` doesn't allow me to override the env vars that it overrides directly. This diff adds an argument that enables this behavior. I don't want to enable this by default to avoid breaking existing workflows, I also don't have a Windows computer to test this with.

Reviewed By: JakobDegen

Differential Revision: D87854609

fbshipit-source-id: 06434920a1a7877292087d965d73555eb6b860e0