Swiftdoc Artifacts
Summary:
# Context
There are certain files that are generated during a Swift compile that are not currently tracked by Buck2. If we do not track them, they won't be available as incremental inputs in RE actions, for instance. So for each, we either need to
1) Not emit the files (sometimes safe, as if they were important output for another action, they would be tracked).
2) Track the files

# This diff

.swiftdoc files are currently not tracked. They do NOT seem to contain absolute paths as far as I can tell (or even relative paths, unless paths have been added via a `/// code comment`). This means we can track them as buck2 artifacts.

We are using output-file-maps, so we can't use ` -emit-module-doc-path` to set the path - it has to be in the output-file-map.

But, when incremental mode is enabled, the "swiftdoc" argument is ignored for the "main" section (i.e. the "" entry).

For this reason, at the moment, we need to rely on the convention that the compiler emits the swiftdoc alongside the swiftmodule.

Reviewed By: rmaz

Differential Revision: D87788795

fbshipit-source-id: 631470a937ee16ad61f0a795f71a7263f949ea53