Don't emit depsfiles with incremental remote
Summary:
# Context
There are certain files that are generated during a Swift compile that are not currently tracked by Buck2. If we do not track them, they won't be available as incremental inputs in RE actions, for instance. So for each, we either need to
1) Not emit the files (sometimes safe, as if they were important output for another action, they would be tracked).
2) Track the files

# This diff

Currently depsfiles are disabled for WA builds using `skip_incremental_outputs`.

They are incompatible with hermeticity for Swift Incremental in RE:
* These files contain absolute paths that are handled by postprocessing:
https://www.internalfb.com/code/fbsource/[0cddcbd4748fdedc40d053fa7e72650426f5d265]/fbcode/buck2/prelude/apple/tools/swift_exec.py?lines=136-142
* However, we must provide the *original* file to enable builds to proceed incrementally. We cannot track these as they contain absolute paths

For this reason, I propose we continue to disable these until we can ensure there are no absolute paths.

Reviewed By: manicaesar

Differential Revision: D87788797

fbshipit-source-id: a647e99dfb10641600b7ec90eb80a9f768223b60