Migrate buck to hyper-http-proxy and newer rustls
Summary: Migrate buck to hyper-http-proxy. This follows on D76918206 (see summary there) by upgrading to a better supported proxy package. This also upgrades us to a newer rustls which is generally better for us long-term. The two upgrades are entangled due to hyper-http-proxy depending on hyper-rustls 0.27 which pulls in rustls 0.23 and tokio-rustls 0.26.

Reviewed By: JakobDegen

Differential Revision: D82972130

fbshipit-source-id: 0ef854f0d5e40b3168127127e68765269755c63f