Remove `cancellation_requested` bool
Summary:
This thing is now just - semantically - fully unnecessary. It suffices to check the state.

Note that this is (hypothetically) a mild perf regression, since this replaces an atomic load with a lock. I wouldn't expect that to show up anywhere, but regardless I'm going to fix that again next diff.

Reviewed By: jtbraun

Differential Revision: D88296650

fbshipit-source-id: cfec585dc6faddf12d92d9a0e8637457f827a255