Unify all the shared state
Summary:
Ok, I know how I want to do this now and I'm confident this is the right path.

We currently have shared state split across a number of different types. With weird amounts of sharing.

This diff gets rid of that and puts it all into `SharedStateData`. Everything holds an `Arc<SharedStateData>`. Next diff will deduplicate some of this data and make it a lot easier to reason about the invariants across it

Reviewed By: jtbraun

Differential Revision: D88296656

fbshipit-source-id: 0c4b6b86de9fa711a60a538bc310a870033d5525