Log schedule type via SANDCASTLE_SCHEDULE_TYPE env var
Summary:
We currently log schedule type based on cache.schedule_type buckconfig, but this is broken because it requires people to set this buckconfig for the schedule type to be applied. In practice, this means schedule_type = user for a lot of CI builds.

The correct way to set schedule type is by reading SANDCASTLE_SCHEDULE_TYPE env var. We already read this env var to do lots of other important things, like not sending cache hit events to ingress for diff-time jobs. This diff updates our logging to use the same variable as well.

Reviewed By: IanChilds

Differential Revision: D88445255

fbshipit-source-id: 2bf3a3eacfc2dc6af68dd85c5a3346f507c88d00