Replace merged_link_infos with tset
Summary:
This greatly reduces memory usage of `rust_library` analysis by tracking link info in a way that does not grow in proportion to the size of the transitive dependency graph of a library. Link info previously accounted for over 60% of the memory usage of Rust library targets.

For `fbcode//zippydb/tools/zu_rs/utils/general:zu_rs_utils`:

||Before|After|Δ|
|--|
|allocations|10,637|10,712|+0.7%|
|allocated bytes|3,529,909|1,376,775|-61.0%|
|retained bytes|3,389,949|1,237,695|-63.5%|

For binaries the effect is less significant because their memory usage is dominated by `rpath-tree` and `shared-libraries` subtargets, which will need a different fix. Once those other things are fixed, the effect of this diff on binary targets will become similar to the effect on libraries.

For the transitive dependency graph of `fbcode//zippydb/tools/zu_rs:zu `:

||Before|After|Δ|
|--|
|allocations|47,506,220|47,814,692|+0.6%|
|allocated bytes|3,565,303,152|3,189,204,740|-10.5%|
|retained bytes|3,448,562,724|3,076,027,328|-10.8%|

Reviewed By: capickett

Differential Revision: D88523218

fbshipit-source-id: 80e4ec6d118f7ed35b17f2783bcb5d3b974ce938