Fix misplaced comment on TransitiveSet iterators
Summary:
From what I have come to understand of the `TransitiveSet` traversal code after working on {D88660249} and {D88660250}, I believe these 4 comments are misplaced. `PreorderTransitiveSetIteratorGen` and friends are *not* similar to a `FilterMap` and are *not* an iterator over values. They are an iterator over all nodes in the transitive set, whether or not the node contains a value.

The thing that is a `FilterMap` is actually `TransitiveSetValuesIteratorGen`. This is an iterator over the values in the transitive set, visiting only nodes that contain values. This is similar to a `FilterMap` that throws out nodes without a value.

Reviewed By: JakobDegen

Differential Revision: D88660248

fbshipit-source-id: 401a6b5a99d925b991425ce9c9efc22625797817