ReActionIdentity: make target and paths optional
Instead of making 'identity' optional in some of the request paths, make the fields inside
ReActionIdentity optional instead.

This enables us to mandate 'identity' for all call RBE client consumer.
We simplify the creation of the identity by introducing a minimal constructor to use in
different places with insufficient data to build a full one.

This change is based on https://github.com/facebook/buck2/pull/1169