cgroups: Remove backoff behavior
Summary:
We have and will have load-bearing depedencies on all of this. For memory reporting, but also soon for scheduling decision. If there are problems with this code, just not doing any of the things this does is not really an option. Ignoring errors and doing our best (this change) is maybe ok, but tbh I think we should be considering hard-failing builds when this stuff breaks.

So delete the backoff thing

Reviewed By: cbnilrem

Differential Revision: D87299085

fbshipit-source-id: 1e576515fb0f3126402553d8f370d1aeb688cc6f