Reduce allocations in cxx_library_parameterized by using deps_all list
Summary:
Reduces allocated memory by re-using `all_deps` list instead of allocating and freeing per call-site.

Re-do of {D88529188}, which was reverted in (D88643664) because of linking errors. This new version creates 2 lists, one for each ordering. I'd like to make the ordering consistent or document why it is the way it is, but that is out of scope for now as the focus of this change is memory improvements.

### Heap Allocated Changes
Metrics from `buck2 profile analysis --recursive --mode heap-allocated --output /tmp/profile fbcode//zippydb/tools/zu_rs:zu`

| Metric | Baseline | After | Delta | Delta (%) |
| --------|------------|--------|-------------|
| Commit | 58b9d2aaddef | 9141f9f828f3 |
| Time(s) | 771.449 | 768.17 | 3 | 0.43% |
| Allocs | 554,679,389 | 552,694,100 | 1,985,289 | 0.36% |
| AllocBytes| 45,973,523,358 | 45,825,951,414 | 147,571,944 | 0.32% |
| list | 113,903,618 | 112,846,736 | 1,056,882 | 0.93% |
| array |  93,120,042 | 92,191,635 | 928,407 | 1.00% |

Other metrics and retained memory have no change.

Reviewed By: dtolnay

Differential Revision: D88860578

fbshipit-source-id: bffa9a7e3eace6dcec95f05ec62da7f3c7771545