test orchestrator: deduplicate ActionKeys in critical path tracking
Summary:
Like it says in the title.

Iterating over all artifacts means you get no deduplication or sharing. This is
very much known to blow out memory (D82625781).

In this particular instance, zdiemer shared this job:

https://www.internalfb.com/sandcastle/job/13510801261863319?entry_point=0&no_redirect=true

When I repro this locally it uses 200+GB of RAM. Can't repro exactly that
commit, because there is some code I don't have access to, but I commented out
the deps I can't access so we don't attempt to use them, then it works:

94c66cb1b73c14889a12fb029fba509dbe0d1b96

A memory profile shows this:

https://fburl.com/scuba/stacks_to_scuba/587qztxp

It's pretty obvious that the memory allocation is a `collect` into a `Vec`, and
this is the most suspicious `collect()` in this stack, so I split it out in a
function and added `#[inline(never)]`.

Then I get this:

https://fburl.com/scuba/stacks_to_scuba/36ro87d5

Note the "total value" in there. It says 302894138395. That's **282 GiB of RAM.**
That is a steep price to pay for a critical path (which you will never see, because
you OOMed).

I then added the intermediary collect into a `HashSet` to dedupe that you see in this diff
and now memory peaks at **28GB**. This sounds like a 10x reduction but it is actually
more, because the baseline memory for the **build** step here is actually about 28GB, so
the change here is 250 -> 0, not 280 -> 30.

Note that I tested this with this buckconfig:

```
[buck2]
  share_action_paths = true
  deduplicate_get_digests_ttl_calls = true
  output_trees_download_semaphore_size = 1073741824
  fingerprint_re_output_trees_eagerly = true
```

I don't know how load bearing that is (but probably the first one is at least a
bit load bearing given what we do here).

Reviewed By: IanChilds

Differential Revision: D88950260

fbshipit-source-id: fcb60264bc5c8f0682b8adbbfe5157235ee8a158