errors: Typed top-level errors
Summary:
Replace the `anyhow` top level errors with typed errors that document the handful of ways that this can actually fail. The errors associated with `draw` operations on components remain `anyhow` in this diff, addressing those is next.

We then use this to improve get some centralized error conversions on the buck2 side

Differential Revision: D88917307

fbshipit-source-id: d8801972e74ac51def3aeb50ae20f162e7684966