Remove sketchy waker reuse
Summary:
This thing ws never realy quite right. Fortunately there is even a truly correct way to achieve similar results, since wakers are comparable for equality

To prove that this check actually gets hit in practice, I added

```
 diff --git a/fbcode/buck2/dice/dice_futures/src/details/cancellable_future.rs b/fbcode/buck2/dice/dice_futures/src/details/cancellable_future.rs
 --- a/fbcode/buck2/dice/dice_futures/src/details/cancellable_future.rs
+++ b/fbcode/buck2/dice/dice_futures/src/details/cancellable_future.rs
@@ -142,6 +142,9 @@
             .as_ref()
             .is_some_and(|w| w.will_wake(cx.waker()))
         {
+            if self.last_waker.as_ref().is_some() {
+                unreachable!("different waker!");
+            }
             self.last_waker = Some(cx.waker().clone());
             self.view.register_waker(cx);
         }
```

And then built fb4a. It didn't panic, meaning we're not actually doing any more work. If there are edge cases where this would actually fire when before it didn't, that means we had an actual bug before in which case it's even more important that we have this.

Note that `will_wake` is just comparing two pointers, so it's quite cheap.

Differential Revision: D88441157

fbshipit-source-id: f8e5960cddf6c61a38afb9e7130108de6384072a