Expand DistInfo with fields related to relocatable resources
Summary:
First attempt in D87862071 had to be reverted, at least two things I could discover sampling some of the test failures:
- the path pointed to in resources.json being a link rather than a direct file reference breaks some pre-existing built-in assumptions
- Windows was not happy :)

As a result, for attempt 2 I'm proposing something closer to the original plan of emitting separate fields for use in packaging tools, while leaving the default as-is. This is not quite as clean as changing the resource DB directly, but the latter will take significantly longer to handle the above cases and I'm not convinced reducing duplication is worth all of that added effort/complexity.

dtolnay's original comment on D87862071 of potentially updating language-specific resource libs to be smarter about this could be interesting to explore as another approach, but is also substantially more work, so I'm hoping we can do this in the interim to at least unblock implicit resources in fbpkg, which continues to be requested.

Reviewed By: dtolnay

Differential Revision: D88796949

fbshipit-source-id: c6e43e28193a868d86a3b009ae03e46d45642373