Pin kotlincd to compiler binary 2.1.0
Summary:
We want to only use one verion of kotlin compiler binary with kotlincd which is currently 2.1.0.
With current setup, users can build with `-c kotlin.kotlin_version=1.8.0` and force kotlincd to use the compiler binary for another version.
This change prevents that. Our aproach is to support non-standard cases with kotlin classic instead which will continue to honour `kotlin.kotlin_version` config.
**In this diff:**
* kotlincd is hardcoded to use kotlin_version=2.1.0
* kotlin bootstrap is hardcoded to use kotlin_version=2.1.0
* kotlin classic honours kotlin.kotlin_version config

Reviewed By: siaojiecai

Differential Revision: D88861759

fbshipit-source-id: 6717bb44741aef51be9966ddce1abd9a42186d9a