Replace exported_link_deps with tset
Summary:
This is the last large collection still used in `RustLinkInfo`. This diff applies the same refactor as D88542325 and previous diffs in this stack: still using `traverse` to process the collection and preserve existing behavior exactly, but not storing the expanded result in every provider, only a tset.

No change in allocated and retained memory (±0.025%).

For `fbcode//zippydb/tools/zu_rs/utils/general:zu_rs_utils`:

||Before|After|Δ|
|--|
|allocations|10,652|10,653|+0.01%|
|allocated bytes|3,426,281|3,426,017|-0.01%|
|retained bytes|3,287,209|3,286,929|-0.01%|

For the transitive dependency graph of `fbcode//zippydb/tools/zu_rs:zu `:

||Before|After|Δ|
|--|
|allocations|47,115,235|47,119,318|+0.01%|
|allocated bytes|3,529,342,021|3,530,211,613|-0.02%|
|retained bytes|3,414,685,161|3,415,494,801|-0.02%|

Reviewed By: capickett

Differential Revision: D88915562

fbshipit-source-id: 3a5d1990a7a0ecce3eaafa864545b7541107ebe3