Simplify dice dump serialized nodes
Summary:
It's no longer true that there can be more than one node per dice key, so merging NodeID and KeyID and renaming nodes -> node.

It's also no longer true that a poisoned lock causes deps or rdeps to be None.

If there are downstream tools reading dice dumps this could break them by changing the format but AFAICT this is only used for manual debugging.

Reviewed By: cjhopman

Differential Revision: D89208308

fbshipit-source-id: 782c2aaa19f6f7c635f0e265f8d9b14f38199794