make buck2 test pass the user to tpx
Summary:
Running `buck2 test` in mobile ODs makes tpx report test runs with user `svcscm`. This is due to the fact that the OD is warmed up as `svcscm` making buck2d being started with that user which in turn will spawn tpx.
To make tpx get the right user also in this case I propose to make buck2 to pass the current user to tpx when invoking it. At that moment `svcscm` should have been replaced with the current user, making tpx reporting be correct.

Reviewed By: asm89

Differential Revision: D89225602

fbshipit-source-id: 7bf7e538396aab1160582348ec496e65e23a7050