Remove bogus concatenations
Summary:
# Context
During listing of Erlang tests, we want to produce a name for each testcase, that include the concatenation of all groups involved, separated by a dot

# Problem
We are currently doing this concatenation explicitly with a foldl. which may be actually quadratic. In any case, this is redundant, as we later process it in a `io_lib:format()` that accepts an `io_list()`.

# This diff
Remove the concatenation and use the `io_list()` instead

Reviewed By: TD5, michalmuskala

Differential Revision: D89375091

fbshipit-source-id: d5c40b8d194f938c2a85e1c512611b0169469af9