add `--chdir` to the `env` in `what-ran` local actions
Summary:
A common use for `buck2 log what-ran` is to use it to find the command line that failed, and then use that to iterate/test/debug what failed. It's convenient if the `env` command also had the `--chdir` argument to put you in the right directly.

N.B. that this is not always going to allow you to reproduce the action. If you're viewing a log for something NOT run on this machine, or you've modified the buck-out state away from when this action was run, you'll get different results. That's always been true, this is just a convenience for when you meet the prerequisites, you can cut/paste the command and run it from any directory (in the buck project, anyway).

Reviewed By: cjhopman

Differential Revision: D89208229

fbshipit-source-id: 1ff324682260c83113a108a7fca96455f5ea2687