fix a flaky test
Summary: Failures to create the cpu instruction counting object are silently ignored when prepping for some starlark evaluations. This can cause the `cpu_instruction_count` field to be `None` in the logs even on linux. We shouldn't let these be actual failures, as it makes for a flaky test.  Convert them to `pytest.xfail()`s so that they are reported as `Skip`'ed tests.

Reviewed By: IanChilds

Differential Revision: D89373697

fbshipit-source-id: d5fc9793ac8396bf11ed15a6bcd586b7f20af993