Fix analysis being done on unconfigured targets
Summary:
Context - https://fb.workplace.com/groups/617497306123691/permalink/1337494374123977/

This bxl script is trying to do analysis with an unconfigured target. Currently this causes a soft error to be thrown - https://fburl.com/scuba/buck2_builds/7hxklhlp. But we should be doing analysis on configured targets. This might turn to a hard error in the future, so we want to migrate such bxl scripts to configure targets before analysis

Reviewed By: IanChilds

Differential Revision: D89437981

fbshipit-source-id: d4dbd62247207f5430051b9a3d61dc5db007912a