make js_bundle_genrule content based path hashed
Summary:
TSIA

Maybe the start of a slow process that looks into why content based path might not bring any concrete wins to igios.

When building igios, there is one action with a category of genrule_js_bundle_genrule, which is super slow, it takes on average >3mins to complete.

If we can dedupe this action, maybe we can get some wins for some users? This diff makes this action eligible for dedupe.

I think it's safe to just enable it by default.

Reviewed By: benb

Differential Revision: D89079455

fbshipit-source-id: 190e5585e1341a925fe83d093264cfc13e443e41