Add Internable trait to associate interned values with interners
Summary:
Supporting pagable for interned values in buck2 is greatly simplified if we can implement it for `Intern<T: Internable>` and be able to access the Interner from T, which is what Internable provides.

The main complication of doing this is that we can't implement Internable for String without wrapping it in a newtype.

The `interner!` macro implements Internable for the interned type, but also supports implementing convenience traits if the interned type is a newtype, to reduce (some of) the friction associated with using a newtype.

Reviewed By: cjhopman

Differential Revision: D89328236

fbshipit-source-id: 00643ad79f63528b8017720ad56877dca5aeb793