Change critical path to use TimeSpan instead of duration
Summary:
Once we can associate critical path entries with timespans, we can
then identify the time between entries and associate those with the following
item. Building on that, we will then provide functionality to further classify
those sections.

This will also allow us to better identify places where we are producing
incorrect data to the critical path calculation, for example it's incorrect for
two things that depend on each other to have overlapping spans.

Reviewed By: JakobDegen

Differential Revision: D84100479

fbshipit-source-id: bf258d4640186afb1de25fd55b9b3339b777192d