buck2: add "audit file-package"
Summary:
Add a straightforward way to map from a file to a Buck package. This
is different from `query owner(...)` because the file does not have
to be a source file - it's just anything contained within the package,
including BUCK or .bzl files.

```
$ buck audit file-package buck2/BUCK ../third-party/rust/Cargo.toml
```

It also accepts the `--json` option for json output.

It takes either project-relative paths, or absolute paths that have a
common prefix with the project root:

```
[jsgf@devvm3302.eag0 /data/users/jsgf/fbsource/fbcode (7c6f3ae848)]$ ./buck2.bin audit file-package $(realpath buck2/BUCK)
/data/users/jsgf/fbsource/fbcode/buck2/BUCK: fbcode//buck2
```
```
[jsgf@devvm3302.eag0 /data/users/jsgf/fbsource/fbcode (7c6f3ae848)]$ ./buck2.bin audit file-package buck2/BUCK
buck2/BUCK: fbcode//buck2
```

Note that symlinks within the repo are not handled specifically; they
probably won't work as expected. Don't use symlinks.

Related: https://fb.workplace.com/groups/askbuck/posts/29873725522249345/

Reviewed By: cjhopman

Differential Revision: D85725494

fbshipit-source-id: 94a6443ce6a265148936f4f582a719ad7c415fcb