rust-project: populate `sysroot_src` from `RUST_SRC_PATH`
According to
https://rust-analyzer.github.io/book/non_cargo_based_projects.html,
rust-analyzer should compute `sysroot_src` from `sysroot` if it is
absent.

According to my testing, this is not the case, and rust-analyzer is
broken if `sysroot_src` is not populated.

To fix this, we populate it from the de facto standard `RUST_SRC_PATH`,
which fixes the issue.

Ideally, rust-analyzer should be fixed but:
1. I haven't gotten to the bottom of that bug.
2. This commit remains correct in the meantime.
3. This commit allows setups where the binary sysroot and source
   sysroots are different (as is the case in Nixpks, where they come
   from different packages).