Improve documentation about build configuration
This commit rearranges existing documentation and adds more explanation,
including a comprehensive overview of the various parts (platforms,
modifiers and transitions), and how they are relevant for end users.

In particular, a lot of information was moved from `rule_authors/` to
`concepts/`, as build configuration is important for end users.

A lot of information was also deduplicated across pages.